SISRP-5405 Chop up CampusSolutionsController into consitutent parts
