+++ app/controllers/campus_solutions/aid_years_controller.rb
@@ -1,0 +1,7 @@
+class AidYearsController < CampusSolutionsController
+
+  def aid_years
+    render json: CampusSolutions::MyAidYears.from_session(session).get_feed_as_json
+  end
+
+end
