+++ app/controllers/campus_solutions/campus_solutions_controller.rb
@@ -1,0 +1,3 @@
+class CampusSolutionsController < ApplicationController
+  before_filter :api_authenticate_401
+end
