+++ app/controllers/campus_solutions/country_controller.rb
@@ -1,0 +1,7 @@
+class CountryController < CampusSolutionsController
+
+  def country
+    render json: CampusSolutions::Country.new.get
+  end
+
+end
