+++ app/controllers/campus_solutions/state_controller.rb
@@ -1,0 +1,7 @@
+class StateController < CampusSolutionsController
+
+  def state
+    render json: CampusSolutions::State.new.get
+  end
+
+end
