+++ app/controllers/campus_solutions/terms_and_conditions_controller.rb
@@ -1,0 +1,8 @@
+class TermsAndConditionsController < CampusSolutionsController
+
+  def terms_and_conditions
+    model = CampusSolutions::MyTermsAndConditions.from_session(session)
+    render json: model.update(request.request_parameters)
+  end
+
+end
