+++ spec/controllers/campus_solutions/title4_controller_spec.rb
@@ -1,0 +1,30 @@
+require 'spec_helper'
+
+describe Title4Controller do
+  context 'updating title 4' do
+    it 'should not let an unauthenticated user post' do
+      post :title4, {format: 'json', uid: '100'}
+      expect(response.status).to eq 401
+    end
+
+    context 'authenticated user' do
+      before do
+        session['user_id'] = '1234'
+        User::Auth.stub(:where).and_return([User::Auth.new(uid: '1234', is_superuser: false, active: true)])
+      end
+      it 'should let an authenticated user post' do
+        post :title4,
+             {
+               bogus_field: 'abc',
+               response: 'Y'
+             }
+        expect(response.status).to eq 200
+        json = JSON.parse(response.body)
+        expect(json['statusCode']).to eq 200
+        expect(json['feed']).to be
+        expect(json['feed']['institution']).to eq 'UCB01'
+      end
+    end
+  end
+end
+
