+++ spec/models/campus_solutions/ethnicity_spec.rb
@@ -17,7 +17,7 @@ describe CampusSolutions::Ethnicity do
     it_should_behave_like 'a proxy that gets data'
   end
 
-  context 'real proxy' do
+  context 'real proxy', testext: true do
     let(:proxy) { CampusSolutions::Ethnicity.new(fake: false) }
     it_should_behave_like 'a proxy that gets data'
   end
