+++ spec/support/campus_solutions_helper_module.rb
@@ -8,4 +8,23 @@ module CampusSolutionsHelperModule
       expect(subject[:errored]).to eq true
     end
   end
+
+  shared_examples 'an unauthenticated user' do
+    it 'returns 401' do
+      get feed
+      expect(response.status).to eq 401
+      expect(response.body.strip).to eq ''
+    end
+  end
+
+  shared_examples 'a successful feed' do
+    it 'has some data' do
+      session['user_id'] = user
+      get feed
+      json = JSON.parse(response.body)
+      expect(json['statusCode']).to eq 200
+      expect(json['feed'][feed_key]).to be
+    end
+  end
+
 end
